Implementing Authentication & Error validation & fix CRUD operation bugs

+++ server/TodoAPI/Controllers/TodoController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Formatters;
 using TodoAPI.DTO;
 using TodoAPI.Models;
 
@@ -8,7 +9,7 @@ namespace TodoAPI.Controllers
 {
     [Route("/[controller]")]
     [ApiController]
-    //[Authorize]
+    //[Authorize("Bearer ")]
     public class TodoController : ControllerBase
     {
         private readonly TodoDBContext context;
@@ -21,7 +22,7 @@ namespace TodoAPI.Controllers
 
         // get all todos based on user Id
         [HttpGet("/user/{id}/todos")]
-        // route: /todo/user/{userId}/todos 
+        // route: /user/{userId}/todos 
         public IActionResult GetAllTodos(int id)
         {
             var todos = context.Todos.Where(t => t.UserId == id).ToList();
@@ -29,10 +30,23 @@ namespace TodoAPI.Controllers
         }
 
 
+        [HttpGet("/todo/{todoId}")]
+        public IActionResult GetSingleTodo (int todoId)
+        {
+            Todo? todo = context.Todos.Where(t => t.TodoId == todoId).SingleOrDefault();
+            if(todo is not null)
+            {
+                return Ok(todo);
+            }
+            else
+            {
+                return BadRequest("Todo not found !!");
+            }
+        }
 
         // Add Todo 
         [HttpPost("/user/{id}/AddTodo")]
-        // route: /todo/user/userId/addTodo
+        // route: /user/{userId}/addTodo
         public IActionResult AddTodo(int id, TodoDTO todoDTO)
         {
             if (!ModelState.IsValid)
@@ -53,21 +67,50 @@ namespace TodoAPI.Controllers
             context.Todos.Add(todo);   
             context.SaveChanges();
 
-            return Ok();
+            return Ok(true);
 
         }
 
 
+        // Update Todo 
+        [HttpPut("/update/todo/{todoId}")]
+        public IActionResult UpdateTodo (int todoId, [FromBody]TodoDTO todoDTO)
+        {
+            Todo? todoToUpdate = context.Todos
+                .Where(t => t.TodoId == todoId)
+                .SingleOrDefault();
+
+            if(todoToUpdate is not null)
+            {
+                todoToUpdate.Title = todoDTO.Title;
+                todoToUpdate.Description = todoDTO.Description;
+                context.Todos.Update(todoToUpdate);
+                context.SaveChanges();
+                return Ok();
+            }
+            return Problem();
+        }
+
+
+            
         [HttpDelete("/todos/{todoId}")]
         public IActionResult DeleteTodo (int todoId)
         {
             try
             {
-                var todoToDelete = context.Todos.Where(t => t.TodoId == todoId).SingleOrDefault();
+                Todo? todoToDelete = context.Todos.Where(t => t.TodoId == todoId).SingleOrDefault();
 
-                context.Todos.Remove(todoToDelete);
-                context.SaveChanges();
-                return Ok();
+                if(todoToDelete is not null)
+                {
+                    context.Todos.Remove(todoToDelete);
+                    context.SaveChanges();
+                    return Ok();
+                }
+                else
+                {
+                    return Problem("Todo not found");
+                }
+
 
             }
             catch
